--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/structure/positions/ClipPositions.java
@@ -1,5 +1,0 @@
-package org.firstinspires.ftc.teamcode.structure.positions;
-
-public enum ClipPositions {
-	OPEN, CLOSE
-}
