--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/structure/positions/LiftMode.java
@@ -1,5 +1,0 @@
-package org.firstinspires.ftc.teamcode.structure.positions;
-
-public enum LiftMode {
-	IDLE, DECANT_LOW, DECANT_HIGH, SUSPEND, SUSPEND_PREPARE, SUSPEND_Lv1, SUSPEND_Lv2_PREPARE, SUSPEND_Lv2
-}
