--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/structure/positions/ScalePositions.java
@@ -1,5 +1,0 @@
-package org.firstinspires.ftc.teamcode.structure.positions;
-
-public enum ScalePositions {
-	BACK, PROBE
-}
